renamed share to common
